Created task-list component which does not function just yet
